#2 - feat: 기능 변경, 검색 엔진을 따로 만들어 사용
카카오맵 검색 API를 쓰지 않고 우리가 따로 만들어 사용하게 되어 검색 기능을 모두 제거 및 원래 파일 보류 폴더에
